divide git_hooks.spect.ts into git_hooks.spec.ts and apply_new_files.spec.ts
